Split bindField method into its seperate trait
