+++ src/Insert.php
@@ -2,10 +2,12 @@
 
 namespace Girgias\QueryBuilder;
 
+use Girgias\QueryBuilder\Traits\BindField;
 use RuntimeException;
 
 class Insert extends Query
 {
+    use BindField;
 
     /**
      * Return built Query
